chore: order creation time range

+++ service/src/main/kotlin/com/openlivery/service/common/repository/ParametersRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ParametersRepository: JpaRepository<Parameters, Long>
+interface ParametersRepository: JpaRepository<Parameters, Short>
